Bridge Low Poly Compressed
There were issues with the glass in the middle of the bridge when compressing. The textures got messed up. Therefor I uploaded the compressed version with the problematic texture (called: GlassProblem) as it is only under the bridge, meaning it can still be used if we use a perspective which will not make it visible. However I also uploaded the compressed bridge without that glass and in a seperate file the low poly NOT compressed version of the glass (with the right texture) so that we can try to place them in the engine seperatly still resulting in lower memory usage.